pyproject.toml
[project]
name = "tp-base-telecom-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "python-dotenv>=1.2.2",
    "scipy>=1.17.1",
    "sounddevice>=0.5.5",
]
